Updated dependencies
- Updated PHP minimum supported version to 5.6.
- Updated all component and module dependencies to the latest stable releases;
  if v3 releases exist, these are supported via `||` constraints.
- Moved `Module` class to `src/` tree, where it belongs, and:
  - Removed unneeded `getAutoloaderConfig()` implementation.
  - Ensured attachment of event listeners is correct.
  - Updated to use short array notation.
  - Added docblocks.
