+++ composer.json
@@ -25,33 +25,38 @@
         }
     },
     "require": {
-        "php": ">=5.5",
-        "zendframework/zend-eventmanager": "~2.3",
-        "zendframework/zend-mvc": "~2.3",
-        "zendframework/zend-servicemanager": "~2.3",
-        "zendframework/zend-view": "~2.3",
-        "zfcampus/zf-api-problem": "~1.0",
+        "php": "^5.6 || ^7.0",
+        "zendframework/zend-eventmanager": "^2.6.3 || ^3.0.1",
+        "zendframework/zend-mvc": "^2.7.9 || ^3.0.2",
+        "zendframework/zend-servicemanager": "^2.7.6 || ^3.1",
+        "zendframework/zend-view": "^2.8.1",
+        "zfcampus/zf-api-problem": "^1.2.1",
         "zfcampus/zf-apigility-documentation": "^1.2",
-        "zfcampus/zf-content-negotiation": "~1.0"
+        "zfcampus/zf-content-negotiation": "^1.2.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.7",
+        "phpunit/phpunit": "^4.7",
         "squizlabs/php_codesniffer": "^2.3.1",
-        "zendframework/zend-http": "~2.3",
-        "zendframework/zend-loader": "~2.3",
-        "zendframework/zend-stdlib": "~2.3"
+        "zendframework/zend-http": "^2.5.4",
+        "zendframework/zend-stdlib": "^2.7.7 || ^3.0.1"
     },
     "autoload": {
         "psr-4": {
             "ZF\\Apigility\\Documentation\\Swagger\\": "src/"
-        },
-        "classmap": [
-            "Module.php"
-        ]
+        }
     },
     "autoload-dev": {
         "psr-4": {
             "ZFTest\\Apigility\\Documentation\\Swagger\\": "test/"
         }
+    },
+    "scripts": {
+        "check": [
+            "@cs-check",
+            "@test"
+        ],
+        "cs-check": "phpcs",
+        "cs-fix": "phpcbf",
+        "test": "phpunit"
     }
 }
